requirements.txt
torch
tqdm
numpy
opencv-python
cv2opencv-contrib-python
